Refactor redux folder structure
