+++ apps/user-portal/src/actions/index.ts
@@ -16,6 +16,8 @@
  * under the License.
  */
 
+export * from "./account-security";
+export * from "./api";
 export * from "./consents";
 export * from "./login";
 export * from "./profile";
